moved landing components to co-located folder
