+++ src/app/[locale]/components/FAQLanding.module.css
