+++ src/app/[locale]/components/HeroScreenshot.module.css
