+++ src/app/[locale]/components/HeroScreenshot.tsx
