+++ src/app/[locale]/components/LandingHero.module.css
