+++ src/app/[locale]/components/LandingHero.tsx
