+++ src/app/[locale]/components/ProblemCarousel.module.css
