+++ src/app/[locale]/components/ProblemCarousel.tsx
@@ -1,4 +1,4 @@
-import Carousel from "../carousel/Carousel";
+import Carousel from "../../../components/carousel/Carousel";
 
 import Problem1Illustration from "../../assets/landing/problem-illustration-devices.svg";
 import Problem2Illustration from "../../assets/landing/problem-illustration-technical.svg";
@@ -7,7 +7,7 @@ import ProblemSquiggle from "../../assets/landing/problem-squiggle.svg";
 import Image from "next/image";
 
 import styles from "./ProblemCarousel.module.css";
-import TitleSubtitle from "../type/TitleSubtitle";
+import TitleSubtitle from "../../../components/type/TitleSubtitle";
 
 export default function ProblemCarousel() {
     const slides = [
