+++ src/app/[locale]/components/Services.tsx
@@ -1,4 +1,4 @@
-import TitleSubtitle from "../type/TitleSubtitle";
+import TitleSubtitle from "../../../components/type/TitleSubtitle";
 import Image, { StaticImageData } from "next/image";
 
 import Service1Illustration from "../../assets/landing/service-image-remotecontrol.jpg";
