+++ src/app/[locale]/components/Testimonials.tsx
@@ -1,6 +1,6 @@
 import Image, { StaticImageData } from "next/image";
-import Marquee from "../marquee/Marquee";
-import TitleSubtitle from "../type/TitleSubtitle";
+import Marquee from "../../../components/marquee/Marquee";
+import TitleSubtitle from "../../../components/type/TitleSubtitle";
 
 import PlaceholderAvatar from "../../assets/landing/placeholder-avatar.jpg";
